implementando o restos dos metodos do DAO e criando o metodo de desqualificar

+++ src/br/com/g101/processoseletivo/service/IdUtils.java
@@ -1,12 +1,11 @@
 package br.com.g101.processoseletivo.service;
 
-import java.util.concurrent.atomic.AtomicReference;
 
 public class IdUtils {
 
-    private static long id = 1;
+    private static int id = 1;
 
-    public static long nextId() {
+    public static int nextId() {
         return id++;
     }
